+++ noodles-sam/src/header/record.rs
@@ -69,18 +69,18 @@ impl From<Record> for (Kind, Value) {
 /// An error returned when a raw SAM header record fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
+    /// The input is invalid.
+    Invalid,
     /// The kind is missing.
     MissingKind,
     /// The kind is invalid.
     InvalidKind(kind::ParseError),
-    /// A tag is missing.
-    MissingTag,
+    /// A field is invalid.
+    InvalidField,
     /// A tag is invalid.
     InvalidTag,
     /// A tag is duplicated.
     DuplicateTag(String),
-    /// A tag value is missing.
-    MissingValue(String),
 }
 
 impl error::Error for ParseError {}
@@ -88,12 +88,12 @@ impl error::Error for ParseError {}
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            Self::Invalid => write!(f, "invalid input"),
             Self::MissingKind => write!(f, "missing kind"),
             Self::InvalidKind(e) => write!(f, "invalid kind: {}", e),
-            Self::MissingTag => write!(f, "missing tag"),
+            Self::InvalidField => write!(f, "invalid field"),
             Self::InvalidTag => write!(f, "invalid tag"),
             Self::DuplicateTag(tag) => write!(f, "duplicate tag: {}", tag),
-            Self::MissingValue(tag) => write!(f, "missing value for tag {}", tag),
         }
     }
 }
@@ -102,56 +102,52 @@ impl FromStr for Record {
     type Err = ParseError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut pieces = s.split(DELIMITER);
+        match s.split_once(DELIMITER) {
+            Some((k, v)) => {
+                let kind = k.parse().map_err(ParseError::InvalidKind)?;
 
-        let kind = pieces
-            .next()
-            .ok_or(ParseError::MissingKind)
-            .and_then(|s| s.parse().map_err(ParseError::InvalidKind))?;
+                let value = match kind {
+                    Kind::Comment => parse_comment(v)?,
+                    _ => parse_map(v)?,
+                };
 
-        let value = match kind {
-            Kind::Comment => parse_comment(&mut pieces)?,
-            _ => parse_map(&mut pieces)?,
-        };
-
-        Ok(Self::new(kind, value))
+                Ok(Self::new(kind, value))
+            }
+            None => Err(ParseError::Invalid),
+        }
     }
 }
 
-fn parse_comment<'a, I>(iter: &mut I) -> Result<Value, ParseError>
-where
-    I: Iterator<Item = &'a str>,
-{
-    iter.next()
-        .map(|s| Value::String(s.into()))
-        .ok_or_else(|| ParseError::MissingValue(Kind::Comment.to_string()))
+fn parse_comment(s: &str) -> Result<Value, ParseError> {
+    Ok(Value::String(s.into()))
 }
 
-fn parse_map<'a, I>(iter: &mut I) -> Result<Value, ParseError>
-where
-    I: Iterator<Item = &'a str>,
-{
+fn parse_map(s: &str) -> Result<Value, ParseError> {
+    let raw_fields = s.split(DELIMITER);
     let mut fields = Fields::new();
 
-    for s in iter {
-        let mut components = s.splitn(2, DATA_FIELD_DELIMITER);
+    for raw_field in raw_fields {
+        let (tag, value) = parse_field(raw_field)?;
 
-        let tag = components.next().ok_or(ParseError::MissingTag)?;
-
-        if !is_valid_tag(tag) {
-            return Err(ParseError::InvalidTag);
-        }
-
-        let value = components
-            .next()
-            .ok_or_else(|| ParseError::MissingValue(tag.into()))?;
-
-        if fields.insert(tag.into(), value.into()).is_some() {
-            return Err(ParseError::DuplicateTag(tag.into()));
+        if fields.insert(tag.clone(), value).is_some() {
+            return Err(ParseError::DuplicateTag(tag));
         }
     }
 
     Ok(Value::Map(fields))
+}
+
+fn parse_field(s: &str) -> Result<(String, String), ParseError> {
+    match s.split_once(DATA_FIELD_DELIMITER) {
+        Some((tag, value)) => {
+            if !is_valid_tag(tag) {
+                return Err(ParseError::InvalidTag);
+            }
+
+            Ok((tag.into(), value.into()))
+        }
+        None => Err(ParseError::InvalidField),
+    }
 }
 
 fn is_valid_tag(s: &str) -> bool {
@@ -228,9 +224,12 @@ mod tests {
         );
 
         assert!(matches!(
-            "@ND".parse::<Record>(),
+            "@ND\t".parse::<Record>(),
             Err(ParseError::InvalidKind(_))
         ));
+
+        assert_eq!("@HD\t".parse::<Record>(), Err(ParseError::InvalidField));
+        assert_eq!("@HD\tVN".parse::<Record>(), Err(ParseError::InvalidField));
 
         assert_eq!("@HD\tV:1.6".parse::<Record>(), Err(ParseError::InvalidTag));
         assert_eq!("@HD\t0V:1.6".parse::<Record>(), Err(ParseError::InvalidTag));
@@ -244,10 +243,7 @@ mod tests {
             Err(ParseError::DuplicateTag(String::from("VN")))
         );
 
-        assert_eq!(
-            "@CO".parse::<Record>(),
-            Err(ParseError::MissingValue(String::from("@CO")))
-        );
+        assert_eq!("@CO".parse::<Record>(), Err(ParseError::Invalid));
 
         Ok(())
     }